Initial commit
You can specifiy all kind of configurations in a 'suites.json' file
and tell which files for which suite will be loaded

Each suite requires a simple .html file with some small bootstrap JS

When the Sprinter repository is in the Specs folder (as a submodule)
you can open the .html file to run the specs directly.

Example:

Base
  - Source
  - Specs
    - Sprinter
    - SpecRunner.html
    - suites.json

Alternatively the Sprinter repository can live as a sibling directory
and can execute multiple .html suites in iframes simultaneously.
with the SpecRunner.html file inside the Sprinter repository.

This way the Sprinter repository does not have to be a submodule.

Example:

Base
  - Source
  - Specs
    - SpecRunner.html
    - suites.json
DOM
  - Source
  - Specs
    - SpecRunner.html
    - suites.json
Sprinter
  - SpecRunner.html

Example of suites.json:

	{
		"SpecRunner": {
			"modules": [
				"Test/Node"
			],
			"requirejs": {
				"paths": {
					"Test": "spec"
				}
			},
			"require": "requirejs"
		}
	}

With the suites.json it will be much easier to create a
spec runner in other environments (common.js / node.js)